# Add `LineIndex` instead of `Pairs::move_cursor` for improve `Pairs::next` and `Pair::line_col` performance.