--- a/grammars/tests/examples.json
+++ b/grammars/tests/examples.json
@@ -26,7 +26,7 @@
     "0123456789": "digit",
     "special": "`1~!@#$%^&*()_+-={':[,]}|;.</>?",
     "hex": "\u0123\u4567\u89AB\uCDEF\uabcd\uef4A",
-    "true": true,
+    "han": "你好", "true": true,
     "false": false,
     "null": null,
     "array":[  ],