Add `LineIndex` instead of `Pairs::move_cursor` for improve Pairs iterate and `Pair::line_col` performance.

Resolve #784
Ref: https://github.com/rust-lang/rust/blob/1.67.0/src/tools/rust-analyzer/crates/ide-db/src/line_index.rs

Benchmark result:

> (Updated on 2022.02.02 18:00) After I fix to for support utf-8, the new `line-index` implementation of line_col benchmark result is close to `fast-line-col`.
> But when In a last file, `line-index` will faster. 

Benchmark file: https://github.com/pest-parser/pest/blob/16de4fa29228ab07a8c49279de6c5f9814496ed6/grammars/benches/json.rs

```
// With 500 times iter
pair.line_col (with LineIndex)               time:   [22.426 µs 22.489 µs 22.556 µs]
position.line_col                            time:   [212.59 µs 213.38 µs 214.29 µs]
position.line_col (with fast-line-col)       time:   [18.241 µs 18.382 µs 18.655 µs]

// With 1000 times iter
pair.line_col (with LineIndex)               time:   [41.160 µs 41.969 µs 43.478 µs]
position.line_col                            time:   [1.7199 ms 1.7246 ms 1.7315 ms]
position.line_col (with fast-line-col)       time:   [48.498 µs 49.450 µs 50.877 µs]

// With 10K times iter
pair.line_col                                time:   [471.75 µs 475.03 µs 478.94 µs]
position.line_col (with fast-line-col)       time:   [8.1941 ms 8.3278 ms 8.5144 ms]
```

And I also tested the cost of `Pairs` iterate.

```
// Before changes
pairs nested iter                    time:   [2.0168 ms 2.0381 ms 2.0725 ms]
pairs flatten iter                   time:   [4.5973 µs 4.6132 µs 4.6307 µs]

// After added LineIndex
pairs nested iter (with LineIndex)   time:   [14.716 µs 14.822 µs 14.964 µs]
pairs flatten iter (with LineIndex)  time:   [5.4637 µs 5.6061 µs 5.7886 µs]
```