@huacnlee 2023-02-01
This change just spend a Vec<usize> memory size that depends on number of lines. Because the LineIndex just storage the line offset, not str.