@huacnlee 2023-02-01
I found it, `Pair#into_inner` not use the `line_index` ref.

After change, new benchmark result:

> And run the new changes in my project, benchmark come to fast.

```
pair.line_col           time:   [10.611 µs 10.651 µs 10.691 µs]
                        change: [+1.3490% +1.9534% +2.4931%] (p = 0.00 < 0.05)
                        Performance has regressed.

position.line_col       time:   [211.16 µs 211.41 µs 211.68 µs]
                        change: [-0.0250% +0.1523% +0.3160%] (p = 0.07 > 0.05)
                        No change in performance detected.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

pairs nested iter       time:   [18.914 µs 19.036 µs 19.216 µs]
                        change: [-99.044% -99.041% -99.037%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

pairs flatten iter      time:   [17.058 µs 17.087 µs 17.119 µs]
                        change: [+0.2530% +0.4559% +0.6712%] (p = 0.00 < 0.05)
                        Change within noise threshold.
```