@CAD97 2023-02-02
This is in fact a reduction in memory cost; the `Rc<LineIndex>` replaces line/col information wherever it's stored, allowing deriving of line/col in `O(col + line)`.

I was and still am hesitant to any changes which require increasing the size of `Pair`, since that impact is multiplied to be roughly linear to input size, but this change is reducing that size cost.

(The point of the cursor/view idea was to not have the line/col information on `Pair`, but apparently that didn't happen?)