@huacnlee 2023-02-02
And I have also test this changes in my project: AutoCorrect

Here is the [benchmark file](https://github.com/huacnlee/autocorrect/blob/v2.5.8/autocorrect/benches/example.rs#L55) test a 2k json file ([Expected ~9ms / iter]((https://github.com/huacnlee/autocorrect/tree/v2.5.0#benchmark)))


## Before use `Position::line_col` in Pest 2.5.2 by enable `fast-line-col`

```
format_json_2k          time:   [19.625 ms 21.815 ms 24.671 ms]
                        change: [+109.28% +132.64% +161.98%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

## After use `Pair::line_col` with `line-index` feature.

```
format_json_2k          time:   [9.3551 ms 9.3770 ms 9.3995 ms]
                        change: [-1.3461% -0.8035% -0.2937%] (p = 0.00 < 0.05)
                        Change within noise threshold.
```

`9.3ms` is same fast like my self implementation of move_cursor in AutoCorrect project.