@huacnlee 2023-02-02
> Shall we remove fast-line-col, since it's not used in line index?

Or we can just deprecated it, so we don't need make a breaking change.

Search code on GitHub, there is looks like not too much used it.

https://github.com/search?q=fast-line-col&type=code