@huacnlee 2023-02-02
Test on https://github.com/kiibo382/prql/blob/main/prql-compiler/benches/bench.rs

```
// Before, Pest 2.5.0
variables-query         time:   [591.85 µs 593.78 µs 595.93 µs]

// After
variables-query         time:   [580.14 µs 581.95 µs 584.15 µs]
```